모듈 분리
