+++ frontend/src/widgets/header/Header.tsx
@@ -3,17 +3,15 @@
 import Link from 'next/link';
 import { useTranslation } from '@/lib/i18n';
 import { useTheme } from '@/lib/theme';
-import { useScrollShadow } from '@/shared/hooks';
 import SunIcon from '@/shared/ui/sunicon';
 import MoonIcon from '@/shared/ui/moonicon';
 
 export default function Header() {
   const { t } = useTranslation();
   const { theme, toggleTheme } = useTheme();
-  const hasScrolled = useScrollShadow({ threshold: 10 });
 
   return (
-    <header className={`w-full bg-background/95 backdrop-blur-md sticky top-0 z-[9999] transition-all duration-300 ${hasScrolled ? 'shadow-md' : ''}`}>
+    <header className="w-full bg-background/95 backdrop-blur-md sticky top-0 z-[9999] transition-all duration-300">
       <div className="max-w-7xl mx-auto px-4 md:px-5 lg:px-6">
         <div className="grid grid-cols-4 md:grid-cols-8 lg:grid-cols-12 gap-4 md:gap-4 lg:gap-6 items-center h-16">
           {/* Logo - span 2 according to ui.md */}
