+++ frontend/src/widgets/home-layout/index.ts
@@ -1,0 +1,1 @@
+export { default as HomeLayout } from './HomeLayout';
